Add logic for getting and removing user specific file names when upload folder doesn't exists.

+++ Classification_Project/FileUtils.py
@@ -62,21 +62,25 @@ class FileUtils:
         return file
 
     def delete_user_files(self, pattern='*'):
-        if pattern == '*':
-            console_logger.info("Clearing tmp folder...")
+        if os.path.exists(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH')):
+            if pattern == '*':
+                console_logger.info("Clearing tmp folder...")
 
-            shutil.rmtree(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
-            os.mkdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
-        else:
-            for file_name in self.get_user_files_names(pattern):
-                os.remove(f'{ApplicationConstants.get_constant("UPLOADS_FOLDER_PATH")}/{file_name}')
+                shutil.rmtree(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
+                os.mkdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
+            else:
+                for file_name in self.get_user_files_names(pattern):
+                    os.remove(f'{ApplicationConstants.get_constant("UPLOADS_FOLDER_PATH")}/{file_name}')
 
     def get_user_files_names(self, pattern='*'):
-        if pattern == '*':
-            return os.listdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
+        if os.path.exists(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH')):
+            if pattern == '*':
+                return os.listdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))
 
-        return list(filter(lambda f: pattern in f,
-                           os.listdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))))
+            return list(filter(lambda f: pattern in f,
+                               os.listdir(ApplicationConstants.get_constant('UPLOADS_FOLDER_PATH'))))
+
+        return []
 
 
 file_utils = FileUtils()
